升级gf与gtoken到最新版本，更新gf生成的model为最新版本

+++ app/model/admin/cms_news/cms_news.go
@@ -49,7 +49,7 @@ type ReqEditParams struct {
 }
 
 //添加文章操作
-func AddNews(req *ReqAddParams, cateIds []int, userId int, tx *gdb.TX) (insId int64, err error) {
+func AddNews(req *ReqAddParams, cateIds []int, userId uint64, tx *gdb.TX) (insId int64, err error) {
 	if len(cateIds) == 0 {
 		err = gerror.New("栏目不能为空")
 		return
